Use prepend instead of alias_method_chain
